moved playable
