+++ src/main/java/org/rookit/dm/album/Album.java
@@ -8,7 +8,7 @@ import java.util.Set;
 
 import org.rookit.dm.artist.Artist;
 import org.rookit.dm.genre.Genreable;
-import org.rookit.dm.playable.Playable;
+import org.rookit.dm.play.Playable;
 import org.rookit.dm.track.Track;
 import org.smof.annnotations.ForceInspection;
 import org.smof.annnotations.SmofIndex;
