+++ src/main/java/org/rookit/dm/play/Playable.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  ******************************************************************************/
-package org.rookit.dm.playable;
+package org.rookit.dm.play;
 
 import java.time.LocalDate;
