+++ src/test/java/org/rookit/dm/utils/TestUtils.java
@@ -7,7 +7,7 @@ import static org.junit.Assert.*;
 
 import org.rookit.dm.genre.Genre;
 import org.rookit.dm.genre.Genreable;
-import org.rookit.dm.playable.Playable;
+import org.rookit.dm.play.Playable;
 import org.rookit.dm.utils.DMTestFactory;
 
 import com.google.common.collect.Sets;
